fixed require statements and script for seed
