+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "horsies",
+  "name": "wings-and-sails",
   "version": "2.0.0",
   "description": "1904 - HappyhorsesTeam! Working on Grace Shopper!",
   "engines": {
